assets/auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Auth authentication microservice"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "asyncpg>=0.29.0",
    "PyJWT[crypto]>=2.8.0",
    "pyotp>=2.9.0",
    "qrcode>=7.4",
    "Pillow>=10.0",
    "python-dotenv>=1.1.1",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "google-auth[requests]>=2.29.0",
]
